// src/utils/open-im-sdk-wasm/types/params.d.ts
export declare type GroupBaseInfo = Partial<Omit<GroupInitInfo, 'groupType'>> & {
    groupID: string;
};